requirements.txt
python-telegram-bot
python-mpv
